requirements.txt
click
rdp
pyyaml